Made this function O(n^2) -> O(n)
Instead of looping through the whole mesh looking for grains that are
connected to the specific node_id and repeating this we instead loop
through the mesh and save all the results to a dictionary that works as
a lookup table.

This means that we only need to loop through all the nodes once instead
of n times.